Re-impelement all the ncurses_functions in Native PHP for php7
PHP 7.0 doesn't have a working ncurses extension yet (There is a patch for it,
but the extension is fairly dead looking and haven't accepted it after nearly
a year).

So, because aparently I hate myself and not wanting to throw away all the nice
ncurses work I did, I re-implemented all of the ncurses_functions that I used
in native PHP to achieve the same end goal. The aim was to have minimal-to-no
changes needed from my existing implementation, which was achieved. (Some minor
bugs were fixed whilst doing it, but generally 0-changes)

The performance is a bit dire at the moment, but i'll look to improve that in
the future, the basic concept is working and lets the VM be viewed!